python-generators-0x00/requirements.txt
mysql-connector-python
dotenv